fix: reorder wake in priority_channel
This is a workaround for what appears to be a more critical bug in
embassy-rp. When testing I observed that the sender side of the channel
would lock up when the senders waker was woken before the queue was
popped. If the mutex is implemented correctly then this should be
impossible.